requirements.txt
flask>=3.0.0
opencv-python>=4.8.0
numpy>=1.24.0
svgwrite>=1.4.3
pyserial>=3.5
Pillow>=10.0.0
openai-whisper>=20231117
